fix: Community API DTO 타입 불일치 및 FETCH JOIN 오류 해결

+++ src/main/java/com/gaebang/backend/domain/community/dto/response/BoardListResponseDto.java
@@ -13,12 +13,12 @@ import java.time.format.DateTimeFormatter;
 public record BoardListResponseDto(
         Long boardId,              // 게시글 ID
         String title,              // 제목
-        int commentCount,          // 댓글 수
+        Long commentCount,          // 댓글 수
         String writer,             // 작성자(글쓴이)
         String imageUrl,           // 이미지 URL
         String createdDate,        // 작성일
-        int viewCount,             // 조회수
-        int likeCount              // 추천수
+        Long viewCount,             // 조회수
+        Long likeCount              // 추천수
 ) {
     /*public static BoardListResponseDto fromEntity(Board board) {
         return new BoardListResponseDto(
